+++ src/compas/geometry/shapes/box.py
@@ -549,7 +549,7 @@ class Box(Shape):
         :class:`compas.brep.Brep`
 
         """
-        from compas.brep import Brep
+        from compas.geometry import Brep
 
         return Brep.from_box(self)