+++ src/compas/geometry/shapes/capsule.py
@@ -342,7 +342,7 @@ class Capsule(Shape):
         :class:`compas.brep.Brep`
 
         """
-        from compas.brep import Brep
+        from compas.geometry import Brep
 
         return Brep.from_capsule(self)