+++ src/compas/geometry/shapes/cone.py
@@ -298,7 +298,7 @@ class Cone(Shape):
         :class:`compas.brep.Brep`
 
         """
-        from compas.brep import Brep
+        from compas.geometry import Brep
 
         return Brep.from_cone(self)