+++ src/compas/geometry/shapes/sphere.py
@@ -264,7 +264,7 @@ class Sphere(Shape):
         :class:`compas.brep.Brep`
 
         """
-        from compas.brep import Brep
+        from compas.geometry import Brep
 
         return Brep.from_sphere(self)