+++ src/compas/geometry/shapes/torus.py
@@ -289,7 +289,7 @@ class Torus(Shape):
         :class:`compas.brep.Brep`
 
         """
-        from compas.brep import Brep
+        from compas.geometry import Brep
 
         return Brep.from_torus(self)