+++ src/compas_rhino/geometry/brep/face.py
@@ -2,7 +2,7 @@ import Rhino
 from Rhino.Geometry import Interval
 from Rhino.Geometry import RevSurface
 
-from compas.brep import BrepFace
+from compas.geometry import BrepFace
 from compas.geometry import Sphere
 from compas.geometry import Cylinder
 from compas.geometry import Frame